ChallengeContent in eigenes Modul ausgelagert
